Module Caqti_blocking_sql_io

include Caqti_sql_io.S with type 'a future = 'a

type 'a future = 'a

val read_sql_statement : ( 'a -> char option future ) -> 'a -> string option future

read_sql_statement read_char chan reads the next semicolon-terminated SQL statement from chan, taking care to skip over quoted semicolons. read_char chan shall return the next character from chan, or None when the end of file has been reached. A final semicolon is optional and any trailing white space after it will be ignored.

This can be used e.g. to read in SQL schemas or schema updates from a file for automatic initialization and updates of tables, sequences, functions, views, etc.

deprecated
New code should instead use Caqti_query.angstrom_parser with the appropriate parsing function from the Angstrom library.