Module Caqti_common_priv

Internal: Prerequisites.

This module is ment for internal use by Caqti and may change in backwards incompatible ways between minor versions without prior notice.

val (%) : ('b -> 'c) -> ('a -> 'b) -> 'a -> 'c

val (%>) : ('a -> 'b) -> ('b -> 'c) -> 'a -> 'c

val (%>?) : ('a -> ('b, 'e) Stdlib.result) -> ('b -> ('c, 'e) Stdlib.result) -> 'a -> ('c, 'e) Stdlib.result

val (|>?) : ('a, 'e) Stdlib.result -> ('a -> ('b, 'e) Stdlib.result) -> ('b, 'e) Stdlib.result

val ident : 'a -> 'a

val ncompose : int -> ('a -> 'a) -> 'a -> 'a

module Option : sig ... end

module Result : sig ... end

module List : sig ... end

val finally : (unit -> unit) -> (unit -> 'a) -> 'a

val datetuple_of_iso8601 : string -> int * int * int

val iso8601_of_datetuple : (int * int * int) -> string

val ptime_of_rfc3339_utc : string -> (Ptime.t, string) Stdlib.result

val pdate_of_iso8601 : string -> (Ptime.t, string) Stdlib.result

val iso8601_of_pdate : Ptime.t -> string

val default_log_src : Logs.Src.t