Module Caqti_connect

Connection functor and backend registration.

val define_loader : ( string -> ( unit, string ) Stdlib.result ) -> unit

Defines the function used to dynamically load driver libraries. This is normally called during initialization by the caqti-dynload library, if linked into the application.

val define_unix_driver : string -> (module Caqti_driver_sig.Of_system_unix) -> unit

define_unix_driver scheme m installs m as a handler for the URI scheme scheme. This call must be done by a backend installed with findlib name caqti-driver-scheme as part of its initialization.

module Make_unix (System : Caqti_driver_sig.System_unix) : Caqti_connect_sig.S with type 'a future := 'a System.future and module Stream = System.Stream

Constructs the main module used to connect to a database for the given concurrency model.