Module Caqti_connect_config

Configuration passed to connect functions.

type _ key

type t

Construction and Lookup

val default : t

The configuration with all keys set to their default values.

val get : 'a key -> t -> 'a

get key cfg is the value associated with key in cfg, which may be a default value if the key has not been explicitly set or if it has been reset.

val set : 'a key -> 'a -> t -> t

set key value cfg associates key with value in cfg.

val reset : 'a key -> t -> t

reset key cfg associates key with its default value in cfg.

Configuration Keys

val tweaks_version : (int * int) key

Declares compatibility with database tweaks introduced up to the given version of Caqti. Defaults to a conservative value.