Module Caqti_connection.Make_populate

Parameters

module Sys : Caqti_driver_sig.System_common

module C : Caqti_connection_sig.Base with type 'a future := 'a Sys.future and type ('a, 'err) stream := ( 'a, 'err ) Sys.Stream.t

Signature

Insertion

val populate : table:string -> columns:string list -> 'a Caqti_type.t -> ( 'a, 'err ) Sys.Stream.t -> ( unit, [> Caqti_error.call_or_retrieve | `Congested of 'err ] ) Stdlib.result Sys.future

populate table columns row_type seq inputs the contents of seq into the database in whatever manner is most efficient as decided by the driver.