Module type Caqti_connection_sig.Populate

type +'a fiber

type (+'a, +'err) stream

Insertion

val populate : table:string -> columns:string list -> 'a Caqti_type.t -> ('a, 'err) stream -> (unit, [> Caqti_error.call_or_retrieve | `Congested of 'err ]) Stdlib.result fiber

populate table columns row_type seq inputs the contents of seq into the database in whatever manner is most efficient as decided by the driver.