Module Caqti_driver_sig

Internal: Signature for driver implementation.

This interface is unstable and may change between minor versions. If you are developing an external driver, please open an issue to sort out requirements and to announce you need for a stable driver API.

module type System_common = sig ... end

module type System_unix = sig ... end

module type S = sig ... end

module type Of_system_unix = functor (System : System_unix) -> S with type 'a future := 'a System.future and type ('a, 'err) stream := ( 'a, 'err ) System.Stream.t