Module Connection_utils.Make_populate

Parameters

module Sys : System_sig.S

module _ : Caqti_connection_sig.Base with type 'a fiber := 'a Sys.Fiber.t and type ('a, 'err) stream := ('a, 'err) Sys.Stream.t

Signature

Insertion

val populate : table:string -> columns:string list -> 'a Caqti_type.t -> ('a, 'err) Sys.Stream.t -> (unit, [> Caqti_error.call_or_retrieve | `Congested of 'err ]) Stdlib.result Sys.Fiber.t

populate table columns row_type seq inputs the contents of seq into the database in whatever manner is most efficient as decided by the driver.