Module Caqti_platform.Connector

Connection functor and backend registration.

val define_loader : (string -> (unit, string) Stdlib.result) -> unit

Defines the function used to dynamically load driver libraries. This is normally called during initialization by the caqti-dynload library, if linked into the application.

val load_library : string -> (unit, string) Stdlib.result

module Make (System : System_sig.S) (Pool : Pool.S with type 'a fiber := 'a System.Fiber.t and type switch := System.Switch.t and type stdenv := System.stdenv) (Loader : Driver_loader.S with type 'a fiber := 'a System.Fiber.t and type switch := System.Switch.t and type stdenv := System.stdenv and type ('a, 'e) stream := ('a, 'e) System.Stream.t) : Caqti_connect_sig.S with type 'a fiber := 'a System.Fiber.t and type ('a, 'e) stream := ('a, 'e) System.Stream.t and type ('a, 'e) pool := ('a, 'e) Pool.t and type 'a with_switch := sw:System.Switch.t -> 'a and type 'a with_stdenv := stdenv:System.stdenv -> 'a and type connection := (module Loader.CONNECTION)

Constructs the main module used to connect to a database for the given concurrency model.