Parameter Make.System

include System_sig.CORE

module Fiber : System_sig.FIBER

type stdenv

Type of an extra argument to connect functions used to pass through the network stack in Mirage and stdenv in EIO. This is eliminated at the service API where not needed.

module Switch : sig ... end

A module used by EIO to handle cleanup tasks; unit for other platforms.

val async : sw:Switch.t -> (unit -> unit Fiber.t) -> unit

async f runs f () asynchroneously if possible, else immediately.

module Semaphore : sig ... end

module Log : sig ... end

module Stream : Caqti_stream_sig.S with type 'a fiber := 'a Fiber.t

module Sequencer : System_sig.SEQUENCER with type 'a fiber := 'a Fiber.t

module Net : System_sig.NET with type 'a fiber := 'a Fiber.t and type switch := Switch.t and type stdenv := stdenv