Parameter Make.Alarm

type t

A handle for cancelling the alarm if supported.

val schedule : sw:System.Switch.t -> stdenv:System.stdenv -> Mtime.t -> (unit -> unit) -> t

If supported, schedule ~sw ~stdenv time f schedules f to be run at time and returns a handle which can be used to unschedule it. The caqti-blocking implementation does nothing. The pool implementation using it makes additional opportunistic calls to the handler. This function must insert a yield before running the function even if the delay is non-positive.

val unschedule : t -> unit

Cancels the alarm if supported. This is only used for early clean-up, so the implementation may choose to let it time out instead.