Module Caqti_platform.System_sig

Signature for concurrency model and OS-calls.

This is the common part of the signature declarning system dependencies. Driver which depend on features from the unix library will also need Caqti_platform_unix.System_sig.S.

module type FIBER = sig ... end

module type SEQUENCER = sig ... end

module type CORE = sig ... end

module type NET = sig ... end

module type S = sig ... end