Module Net.Socket

A socket with input and output channels and dedicated IO functions. This bundling is done to support the various APIs involved for networking and StartTLS.

type t

val output_char : t -> char -> unit Fiber.t

val output_string : t -> string -> unit Fiber.t

val flush : t -> unit Fiber.t

val input_char : t -> char Fiber.t

val really_input : t -> Stdlib.Bytes.t -> int -> int -> unit Fiber.t

val close : t -> unit Fiber.t