Module type Caqti_pool_sig.S

type +'a future

type ('a, +'e) t

val create : ?⁠max_size:int -> ?⁠max_idle_size:int -> ?⁠check:('a -> (bool -> unit) -> unit) -> ?⁠validate:('a -> bool future) -> (unit -> ('a, 'e) Stdlib.result future) -> ('a -> unit future) -> ('a, 'e) t

Internal: create alloc free is a pool of resources allocated by alloc and freed by free. This is primarily intended for implementing the connect_pool functions.

parameter max_size

The maximum number of allocated resources.

parameter max_idle_size

The maximum number of resources to pool for later use. Defaults to max_size.

parameter check

A function used to check a resource after use.

parameter validate

A function to check before use that a resource is still valid.

val size : ('a, 'e) t -> int

size pool is the current number of open resources in pool.

val use : ?⁠priority:float -> ('a -> ('b, 'e) Stdlib.result future) -> ('a, 'e) t -> ('b, 'e) Stdlib.result future

use f pool calls f on a resource drawn from pool, handing back the resource to the pool when f exits.

parameter priority

Requests for the resource are handled in decreasing order of priority. The default priority is 0.0.

val drain : ('a, 'e) t -> unit future

drain pool closes all resources in pool. The pool is still usable, as new resources will be created on demand.