Module Caqti_sql

SQL utility functions.

Note. Since real databases generally do not implement the precise same escaping mechanisms, and discrepancies between the escape function and the database can lead to SQL injection, the standard SQL escaping functions provided below are hardly useful for real applications.

val bprint_sql_escaped : Stdlib.Buffer.t -> string -> unit

bprint_sql_escaped buf s adds an s to buf with any single quoted doubled according to the standard SQL escaping rule.

deprecated
See note in module introduction.

val bprint_sql_quoted : Stdlib.Buffer.t -> string -> unit

bprint_sql_quoted buf s adds s to buf as an SQL-quoted string. Same as bprint_sql_escaped, except surrounded by single quotes.

deprecated
See note in module introduction.

val sql_escaped : string -> string

sql_escaped s is s with single-quotes doubled according to the standard SQL escaping rule.

deprecated
See note in module introduction.

val sql_quoted : string -> string

sql_quoted s is sql_escaped s between a pair of single quotes.

deprecated
See note in module introduction.