Module Caqti_driver_lib

Internal: Library for Drivers

val linear_param_length : Caqti_query.t -> int

linear_param_length templ is the number of linear parameters expected by a query represented by templ.

val linear_param_order : Caqti_query.t -> int list list

linear_param_order templ is a list where item number i is a list of positions of the linearized query which refer to the ith incoming parameter. Positions are zero-based.

val linear_query_string : Caqti_query.t -> string

linear_query_string templ is templ where "?" is substituted for parameters.