Module Caqti_heap

Internal: Min-heap.

This is a very simple implementation deemed sufficient for Caqti_pool. There are more optimal algorithms if you need the best performance on larger heaps.

module type S = sig ... end

module Make : functor (Elt : Stdlib.Set.OrderedType) -> S with type elt = Elt.t